+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
  - DialingPlan configuration is now a single path
+ - Added VolumePoints and Partitions to the room console
 
 ## [10.0.1] - 2019-01-16
 ### Changed